requirements.txt
python-kasa==0.4.0.dev2